this is working much better.
removed post_ok as there are cases when it wouldn't get set properly
and everything would block. The chirping sensor will never have a long
coap timeout to deal with anyway (since it's using NON)

Have the DHT callback return bad if the checksum fails. This was also
another reason for everything to block forever.